Update canvas.js
Rectangles can be drawn by clicking, dragging, and releasing the mouse. When the mouse is released the finalized panel is pushed to the panel array. Whenever a panel is updated all other panels are redrawn using information from the panel array. When the window is resized the entire canvas is also redrawn